+++ NEWS.md
@@ -1,4 +1,4 @@
-# biocdownloads 0.1.0
+# biocpkgs 0.1.0
 
 - First release: `pkg_download_data` function <2017-06-27 Tue>
 - Add function for package dependencies <2017-06-30 Fri>